fix or silence various harmless compiler warnings (msvc), mostly about unused variables or safe narrowing conversions
(cherry picked from commit 54550490c79d81d0353e666a554d13550c16ceb2)

+++ test/test_priority_queue.cpp
@@ -50,7 +50,7 @@ int test_main( int /* argc */, char* /* argv */[] )
     }
     BOOST_CHECK(a_priority_queue.size() == a_priority_queue1.size());
     
-    for(int i = a_priority_queue.size(); i-- > 0;){
+    for(size_t i = a_priority_queue.size(); i-- > 0;){
         const A & a1 = a_priority_queue.top();
         const A & a2 = a_priority_queue1.top();
         BOOST_CHECK(a1 == a2);
